Extract TestSetup module
