+++ services/galley/test/integration/TestSetup.hs
@@ -1,0 +1,51 @@
+module TestSetup
+  ( test
+  , TestSignature
+  , TestSetup
+  , manager
+  , galley
+  , brig
+  , cannon
+  , awsEnv
+  , maxConvSize
+  , Galley
+  , Brig
+  , Cannon
+  , ResponseLBS
+  ) where
+
+import Imports
+import Test.Tasty
+import Test.Tasty.HUnit
+import Control.Lens
+
+import Bilge (Request, Http, Manager, Response)
+import qualified Galley.Aws             as Aws
+
+type Galley      = Request -> Request
+type Brig        = Request -> Request
+type Cannon      = Request -> Request
+type ResponseLBS = Response (Maybe LByteString)
+
+
+type TestSignature a = Galley -> Brig -> Cannon -> Maybe Aws.Env -> Http a
+
+data TestSetup = TestSetup
+  { _manager         :: Manager
+  , _galley          :: Galley
+  , _brig            :: Brig
+  , _cannon          :: Cannon
+  , _awsEnv          :: Maybe Aws.Env
+  , _maxConvSize     :: Word16
+  }
+
+makeLenses ''TestSetup
+
+
+test :: IO TestSetup -> TestName -> TestSignature a -> TestTree
+test s n h = testCase n runTest
+  where
+    runTest = do
+        setup <- s
+        void $ runHttpT (manager setup) (h (galley setup) (brig setup) (cannon setup) (awsEnv setup))
+
